+++ MessageApp/shared-lib/src/main/java/org/example/BulletinBoard.java
@@ -4,6 +4,6 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface BulletinBoard extends Remote {
-    void add(int idx, byte[] value, String tag) throws RemoteException;
-    Pair get(int idx, String preimage) throws RemoteException;
+    void add(long idx, byte[] value, String tag) throws RemoteException;
+    Pair get(long idx, String preimage) throws RemoteException;
 }
